+++ src/main/groovy/com/cdancy/gradle/artifactory/rest/ArtifactoryRestPlugin.groovy
@@ -27,7 +27,7 @@ import org.gradle.api.artifacts.Configuration
  */
 class ArtifactoryRestPlugin implements Plugin<Project> {
     public static final String ARTIFACTORY_CONFIGURATION_NAME = 'artifactoryRest'
-    public static final String ARTIFACTORY_REST_DEFAULT_VERSION = '0.0.3'
+    public static final String ARTIFACTORY_REST_DEFAULT_VERSION = '0.0.4'
     public static final String EXTENSION_NAME = 'artifactoryRest'
     public static final String DEFAULT_TASK_GROUP = 'ArtifactoryRest'
 
@@ -51,11 +51,13 @@ class ArtifactoryRestPlugin implements Plugin<Project> {
         }
     }
 
-    private static Set<File> configurePluginClassPath(Project project) {
+    private static Configuration configurePluginClassPath(Project project) {
         project.repositories.addAll(project.buildscript.repositories.collect())
-        project.configurations.getByName(ARTIFACTORY_CONFIGURATION_NAME).defaultDependencies { dependencies ->
+        Configuration configuration = project.configurations.getByName(ARTIFACTORY_CONFIGURATION_NAME)
+        configuration.defaultDependencies { dependencies ->
             dependencies.add(project.dependencies.create("com.cdancy:artifactory-rest:$ArtifactoryRestPlugin.ARTIFACTORY_REST_DEFAULT_VERSION"))
             dependencies.add(project.dependencies.create('org.slf4j:slf4j-simple:1.7.5'))
-        }.files
+        }
+        configuration
     }
 }