+++ RELEASE_NOTES.md
@@ -1,12 +1,15 @@
-### Version 0.0.4 (TBA)
+### Version 0.0.4 (6/29/16)
 * ADDED: `DeleteArtifact` gained optional property `failFast`.
+* ADDED: task `LatestVersionFromLayout`.
+* REFACTOR: bump version of gradle to 2.13
+* REFACTOR: bump version of artifactory-rest to 0.0.4
 
 ### Version 0.0.3 (April 10, 2016)
 * ADDED: bumped to version 0.0.3 of artifactory-rest. 
 * ADDED: bumped to version 2.12 for gradlew wrapper.
 
 ### Version 0.0.2 (April 6, 2016)
-* ADDED: Task CopyArtifact 
+* ADDED: Task CopyArtifact.
 
 ### Version 0.0.1 (March 20, 2016)
-* init for project
+* init for project.